Removed intake run motor (uses magazine)
Updated intake commands
